requirements.txt
django>=4.1.5,<=5.0.0
djangorestframework
pyyaml
requests
django-cors-headers